requirements.txt
pytest==7.4.3
pytest-html==4.1.1
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0
pytest-asyncio==0.23.2
requests==2.31.0
aiohttp==3.9.1
docker==6.1.3
kubernetes==28.1.0
redis==5.0.1
psycopg2-binary==2.9.9
pymongo==4.6.0
pyyaml==6.0.1
faker==20.1.0
parameterized==0.9.0
jsonschema==4.20.0
cryptography==41.0.8
prometheus-client==0.19.0 